Move extension datasources, resources and schemas into a new ext folder
