+++ internal/ext/decix/resources/cloud_router_network_service_config_cloud_vc.go
@@ -7,17 +7,19 @@ import (
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/crud"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/ixapi"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/schemas"
+	decixschemas "github.com/ix-api-net/terraform-provider-ixapi/internal/ext/decix/schemas"
+	coreresources "github.com/ix-api-net/terraform-provider-ixapi/internal/resources"
 )
 
-func NewCloudRouterNetworkServiceConfigP2PVCResource() *schema.Resource {
+func NewCloudRouterNetworkServiceConfigCloudVCResource() *schema.Resource {
 	return &schema.Resource{
-		Description:   "Use the `ixapi_de_cix_cloud_router_network_service_config_p2p_vc` resource to configure a Cloud ROUTER connection to a point-to-point virtual circuit network service.",
-		CreateContext: crud.Create(cloudRouterConfigP2PVCCreate),
-		UpdateContext: crud.Update(cloudRouterConfigP2PVCUpdate),
-		ReadContext:   crud.Read(cloudRouterConfigP2PVCRead),
-		DeleteContext: crud.Delete(cloudRouterConfigP2PVCDelete),
+		Description:   "Use the `ixapi_de_cix_cloud_router_network_service_config_cloud_vc` resource to configure a Cloud ROUTER connection to a cloud virtual circuit network service.",
+		CreateContext: crud.Create(cloudRouterConfigCloudVCCreate),
+		UpdateContext: crud.Update(cloudRouterConfigCloudVCUpdate),
+		ReadContext:   crud.Read(cloudRouterConfigCloudVCRead),
+		DeleteContext: crud.Delete(cloudRouterConfigCloudVCDelete),
 
-		Schema: schemas.CloudRouterNetworkServiceConfigP2PVCSchema(),
+		Schema: decixschemas.CloudRouterNetworkServiceConfigCloudVCSchema(),
 
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
@@ -25,17 +27,17 @@ func NewCloudRouterNetworkServiceConfigP2PVCResource() *schema.Resource {
 	}
 }
 
-func cloudRouterConfigP2PVCRequestFromResourceData(
+func cloudRouterConfigCloudVCRequestFromResourceData(
 	r *schema.ResourceData,
 ) (*ixapi.CloudRouterNetworkServiceConfigRequest, error) {
-	vlanConfig, err := vlanConfigFromResourceData(r)
+	vlanConfig, err := coreresources.VlanConfigFromResourceData(r)
 	if err != nil {
 		return nil, err
 	}
 
 	res := schemas.ResourceDataFrom(r)
 	req := &ixapi.CloudRouterNetworkServiceConfigRequest{
-		Type:                  "p2p_vc",
+		Type:                  "cloud_vc",
 		ManagingAccount:       res.GetString("managing_account"),
 		BillingAccount:        res.GetString("billing_account"),
 		ConsumingAccount:      res.GetString("consuming_account"),
@@ -51,16 +53,16 @@ func cloudRouterConfigP2PVCRequestFromResourceData(
 		PolicyEgress:          res.GetStringOpt("policy_egress"),
 		AdminStatus:           res.GetString("admin_status"),
 		BFDEnabled:            res.GetBool("bfd_enabled"),
-		CloudVLAN:             nil,
-		Handover:              nil,
-		Connection:            res.GetStringOpt("network_connection"),
+		CloudVLAN:             res.GetIntOpt("cloud_vlan"),
+		Handover:              res.GetIntOpt("handover"),
+		Connection:            nil,
 		PurchaseOrder:         res.GetStringOpt("purchase_order"),
 		NetworkFeatureConfigs: res.GetStringList("network_feature_configs"),
 	}
 	return req, nil
 }
 
-func cloudRouterConfigP2PVCPatchFromResourceData(
+func cloudRouterConfigCloudVCPatchFromResourceData(
 	r *schema.ResourceData,
 ) *ixapi.CloudRouterNetworkServiceConfigPatch {
 	res := schemas.ResourceDataFrom(r)
@@ -71,7 +73,7 @@ func cloudRouterConfigP2PVCPatchFromResourceData(
 	}
 }
 
-func cloudRouterConfigP2PVCCreate(
+func cloudRouterConfigCloudVCCreate(
 	ctx context.Context,
 	res *schema.ResourceData,
 	api *ixapi.Client,
@@ -80,7 +82,7 @@ func cloudRouterConfigP2PVCCreate(
 		return err
 	}
 
-	req, err := cloudRouterConfigP2PVCRequestFromResourceData(res)
+	req, err := cloudRouterConfigCloudVCRequestFromResourceData(res)
 	if err != nil {
 		return err
 	}
@@ -91,10 +93,10 @@ func cloudRouterConfigP2PVCCreate(
 	}
 	res.SetId(config.ID)
 
-	return cloudRouterConfigP2PVCRead(ctx, res, api)
+	return cloudRouterConfigCloudVCRead(ctx, res, api)
 }
 
-func cloudRouterConfigP2PVCRead(
+func cloudRouterConfigCloudVCRead(
 	ctx context.Context,
 	res *schema.ResourceData,
 	api *ixapi.Client,
@@ -119,7 +121,7 @@ func cloudRouterConfigP2PVCRead(
 	return nil
 }
 
-func cloudRouterConfigP2PVCUpdate(
+func cloudRouterConfigCloudVCUpdate(
 	ctx context.Context,
 	res *schema.ResourceData,
 	api *ixapi.Client,
@@ -129,14 +131,14 @@ func cloudRouterConfigP2PVCUpdate(
 	}
 
 	id := res.Id()
-	patch := cloudRouterConfigP2PVCPatchFromResourceData(res)
+	patch := cloudRouterConfigCloudVCPatchFromResourceData(res)
 	if _, err := api.CloudRouterNetworkServiceConfigsPatch(ctx, id, patch); err != nil {
 		return err
 	}
-	return cloudRouterConfigP2PVCRead(ctx, res, api)
+	return cloudRouterConfigCloudVCRead(ctx, res, api)
 }
 
-func cloudRouterConfigP2PVCDelete(
+func cloudRouterConfigCloudVCDelete(
 	ctx context.Context,
 	res *schema.ResourceData,
 	api *ixapi.Client,
@@ -167,5 +169,5 @@ func cloudRouterConfigP2PVCDelete(
 	if err != nil && !ixapi.IsErrNotFound(err) {
 		return err
 	}
-	return cloudRouterConfigP2PVCRead(ctx, res, api)
+	return cloudRouterConfigCloudVCRead(ctx, res, api)
 }
