+++ internal/ext/decix/resources/cloud_router_network_service_config_cloud_vc_test.go
