+++ internal/ext/decix/resources/cloud_router_network_service_config_p2p_vc_test.go
