+++ internal/ext/decix/resources/cloud_router_static_route.go
@@ -7,6 +7,7 @@ import (
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/crud"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/ixapi"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/schemas"
+	decixschemas "github.com/ix-api-net/terraform-provider-ixapi/internal/ext/decix/schemas"
 )
 
 func NewCloudRouterStaticRouteResource() *schema.Resource {
@@ -17,7 +18,7 @@ func NewCloudRouterStaticRouteResource() *schema.Resource {
 		ReadContext:   crud.Read(staticRouteRead),
 		DeleteContext: crud.Delete(staticRouteDelete),
 
-		Schema: schemas.StaticRouteSchema(),
+		Schema: decixschemas.StaticRouteSchema(),
 
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
