+++ internal/provider/provider.go
@@ -8,6 +8,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/datasources"
+	decixds "github.com/ix-api-net/terraform-provider-ixapi/internal/ext/decix/datasources"
+	decixrs "github.com/ix-api-net/terraform-provider-ixapi/internal/ext/decix/resources"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/ixapi"
 	"github.com/ix-api-net/terraform-provider-ixapi/internal/resources"
 )
@@ -108,24 +110,24 @@ func New(version string) func() *schema.Provider {
 			DataSourcesMap: map[string]*schema.Resource{
 				"ixapi_accounts":             datasources.NewAccountsDataSource(),
 				"ixapi_account":              datasources.NewAccountDataSource(),
-				"ixapi_de_cix_cloud_routers":        datasources.NewCloudRoutersDataSource(),
-				"ixapi_de_cix_cloud_router":         datasources.NewCloudRouterDataSource(),
-				"ixapi_de_cix_cloud_router_network_service_configs_cloud_vc": datasources.NewCloudRouterNetworkServiceConfigsCloudVCDataSource(),
-				"ixapi_de_cix_cloud_router_network_service_config_cloud_vc":  datasources.NewCloudRouterNetworkServiceConfigCloudVCDataSource(),
-				"ixapi_de_cix_cloud_router_network_service_configs_p2p_vc":   datasources.NewCloudRouterNetworkServiceConfigsP2PVCDataSource(),
-				"ixapi_de_cix_cloud_router_network_service_config_p2p_vc":    datasources.NewCloudRouterNetworkServiceConfigP2PVCDataSource(),
-				"ixapi_de_cix_cloud_router_network_service_config_advertised_routes": datasources.NewCloudRouterNetworkServiceConfigAdvertisedRoutesDataSource(),
-				"ixapi_de_cix_cloud_router_network_service_config_received_routes":  datasources.NewCloudRouterNetworkServiceConfigReceivedRoutesDataSource(),
-				"ixapi_de_cix_cloud_router_routes":                                    datasources.NewCloudRouterRoutesDataSource(),
-				"ixapi_de_cix_cloud_router_bgp_state":                                 datasources.NewCloudRouterBGPStateDataSource(),
-				"ixapi_de_cix_cloud_router_bfd_state":                        datasources.NewCloudRouterBFDStateDataSource(),
-				"ixapi_de_cix_cloud_router_prefix_lists":                     datasources.NewCloudRouterPrefixListsDataSource(),
-				"ixapi_de_cix_cloud_router_prefix_list":                      datasources.NewCloudRouterPrefixListDataSource(),
-				"ixapi_de_cix_cloud_router_policies":                         datasources.NewCloudRouterPoliciesDataSource(),
-				"ixapi_de_cix_cloud_router_policy":                           datasources.NewCloudRouterPolicyDataSource(),
-				"ixapi_de_cix_cloud_router_static_routes":                    datasources.NewCloudRouterStaticRoutesDataSource(),
-				"ixapi_de_cix_cloud_router_static_route":                     datasources.NewCloudRouterStaticRouteDataSource(),
-				"ixapi_de_cix_cloud_router_arp_table":                        datasources.NewCloudRouterArpTableDataSource(),
+				"ixapi_de_cix_cloud_routers":        decixds.NewCloudRoutersDataSource(),
+				"ixapi_de_cix_cloud_router":         decixds.NewCloudRouterDataSource(),
+				"ixapi_de_cix_cloud_router_network_service_configs_cloud_vc": decixds.NewCloudRouterNetworkServiceConfigsCloudVCDataSource(),
+				"ixapi_de_cix_cloud_router_network_service_config_cloud_vc":  decixds.NewCloudRouterNetworkServiceConfigCloudVCDataSource(),
+				"ixapi_de_cix_cloud_router_network_service_configs_p2p_vc":   decixds.NewCloudRouterNetworkServiceConfigsP2PVCDataSource(),
+				"ixapi_de_cix_cloud_router_network_service_config_p2p_vc":    decixds.NewCloudRouterNetworkServiceConfigP2PVCDataSource(),
+				"ixapi_de_cix_cloud_router_network_service_config_advertised_routes": decixds.NewCloudRouterNetworkServiceConfigAdvertisedRoutesDataSource(),
+				"ixapi_de_cix_cloud_router_network_service_config_received_routes":  decixds.NewCloudRouterNetworkServiceConfigReceivedRoutesDataSource(),
+				"ixapi_de_cix_cloud_router_routes":                                    decixds.NewCloudRouterRoutesDataSource(),
+				"ixapi_de_cix_cloud_router_bgp_state":                                 decixds.NewCloudRouterBGPStateDataSource(),
+				"ixapi_de_cix_cloud_router_bfd_state":                        decixds.NewCloudRouterBFDStateDataSource(),
+				"ixapi_de_cix_cloud_router_prefix_lists":                     decixds.NewCloudRouterPrefixListsDataSource(),
+				"ixapi_de_cix_cloud_router_prefix_list":                      decixds.NewCloudRouterPrefixListDataSource(),
+				"ixapi_de_cix_cloud_router_policies":                         decixds.NewCloudRouterPoliciesDataSource(),
+				"ixapi_de_cix_cloud_router_policy":                           decixds.NewCloudRouterPolicyDataSource(),
+				"ixapi_de_cix_cloud_router_static_routes":                    decixds.NewCloudRouterStaticRoutesDataSource(),
+				"ixapi_de_cix_cloud_router_static_route":                     decixds.NewCloudRouterStaticRouteDataSource(),
+				"ixapi_de_cix_cloud_router_arp_table":                        decixds.NewCloudRouterArpTableDataSource(),
 				"ixapi_connections":                                   datasources.NewConnectionsDataSource(),
 				"ixapi_connection":           datasources.NewConnectionDataSource(),
 				"ixapi_contacts":             datasources.NewContactsDataSource(),
@@ -181,18 +183,18 @@ func New(version string) func() *schema.Provider {
 				"ixapi_product_offering_mp2mp_vc":      datasources.NewProductOfferingMP2MPVCDataSource(),
 				"ixapi_product_offerings_cloud_vc":     datasources.NewProductOfferingsCloudVCDataSource(),
 				"ixapi_product_offering_cloud_vc":      datasources.NewProductOfferingCloudVCDataSource(),
-				"ixapi_de_cix_product_offerings_cloud_vrf": datasources.NewProductOfferingsCloudVRFDataSource(),
-				"ixapi_de_cix_product_offering_cloud_vrf":  datasources.NewProductOfferingCloudVRFDataSource(),
+				"ixapi_de_cix_product_offerings_cloud_vrf": decixds.NewProductOfferingsCloudVRFDataSource(),
+				"ixapi_de_cix_product_offering_cloud_vrf":  decixds.NewProductOfferingCloudVRFDataSource(),
 			},
 			ResourcesMap: map[string]*schema.Resource{
 				"ixapi_contact":          resources.NewContactResource(),
 				"ixapi_account":          resources.NewAccountResource(),
-				"ixapi_de_cix_cloud_router":     resources.NewCloudRouterResource(),
-				"ixapi_de_cix_cloud_router_network_service_config_cloud_vc": resources.NewCloudRouterNetworkServiceConfigCloudVCResource(),
-				"ixapi_de_cix_cloud_router_network_service_config_p2p_vc":   resources.NewCloudRouterNetworkServiceConfigP2PVCResource(),
-				"ixapi_de_cix_cloud_router_prefix_list":            resources.NewCloudRouterPrefixListResource(),
-				"ixapi_de_cix_cloud_router_policy":                 resources.NewCloudRouterPolicyResource(),
-				"ixapi_de_cix_cloud_router_static_route":           resources.NewCloudRouterStaticRouteResource(),
+				"ixapi_de_cix_cloud_router":     decixrs.NewCloudRouterResource(),
+				"ixapi_de_cix_cloud_router_network_service_config_cloud_vc": decixrs.NewCloudRouterNetworkServiceConfigCloudVCResource(),
+				"ixapi_de_cix_cloud_router_network_service_config_p2p_vc":   decixrs.NewCloudRouterNetworkServiceConfigP2PVCResource(),
+				"ixapi_de_cix_cloud_router_prefix_list":            decixrs.NewCloudRouterPrefixListResource(),
+				"ixapi_de_cix_cloud_router_policy":                 decixrs.NewCloudRouterPolicyResource(),
+				"ixapi_de_cix_cloud_router_static_route":           decixrs.NewCloudRouterStaticRouteResource(),
 				"ixapi_mac":              resources.NewMACResource(),
 				"ixapi_connection":       resources.NewConnectionResource(),
 				"ixapi_port_reservation": resources.NewPortReservationResource(),
